+++ src/client/App.tsx
@@ -5,6 +5,7 @@ import { ObjektePage } from './pages/ObjektePage'
 import { VertraegePage } from './pages/VertraegePage'
 import { FinanzenPage } from './pages/FinanzenPage'
 import { ZaehlerPage } from './pages/ZaehlerPage'
+import { NebenkostenPage } from './pages/NebenkostenPage'
 
 export default function App() {
   return (
@@ -17,6 +18,8 @@ export default function App() {
         <Route path="/vertraege/:id" element={<VertraegePage />} />
         <Route path="/finanzen" element={<FinanzenPage />} />
         <Route path="/finanzen/:tab" element={<FinanzenPage />} />
+        <Route path="/nebenkosten" element={<NebenkostenPage />} />
+        <Route path="/nebenkosten/:tab" element={<NebenkostenPage />} />
         <Route path="/nebenkosten/zaehler" element={<ZaehlerPage />} />
         {/* Weitere Routes werden config-gesteuert hinzugefügt */}
       </Routes>